requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
av=8.0.3=pypi_0
black=21.4b2=pypi_0
ca-certificates=2021.4.13=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py37h06a4308_0
chardet=4.0.0=pypi_0
click=8.0.1=pypi_0
cloudpickle=1.6.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.23=pypi_0
detectron2=0.4=dev_0
future=0.18.2=pypi_0
fvcore=0.1.5=pypi_0
google-auth=1.30.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.37.1=pypi_0
hydra-core=1.1.0rc1=pypi_0
idna=2.10=pypi_0
importlib-metadata=4.0.1=pypi_0
importlib-resources=5.1.3=pypi_0
iopath=0.1.7=pypi_0
joblib=1.0.1=pypi_0
kiwisolver=1.3.1=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.3.4=pypi_0
matplotlib=3.4.2=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.20.3=pypi_0
oauthlib=3.1.0=pypi_0
omegaconf=2.1.0rc1=pypi_0
opencv-python=4.5.2.52=pypi_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.4=pypi_0
pathspec=0.8.1=pypi_0
pillow=8.2.0=pypi_0
pip=21.0.1=py37h06a4308_0
portalocker=2.3.0=pypi_0
protobuf=3.17.0=pypi_0
psutil=5.8.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pydot=1.4.2=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
regex=2021.4.4=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.6.3=pypi_0
setuptools=52.0.0=py37h06a4308_0
simplejson=3.17.2=pypi_0
six=1.16.0=pypi_0
sklearn=0.0=pypi_0
slowfast=1.0=dev_0
sqlite=3.35.4=hdfb4753_0
tabulate=0.8.9=pypi_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
torch=1.8.0+cu111=pypi_0
torchaudio=0.8.0=pypi_0
torchvision=0.9.0+cu111=pypi_0
tqdm=4.60.0=pypi_0
typed-ast=1.4.3=pypi_0
typing-extensions=3.10.0.0=pypi_0
urllib3=1.26.4=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
